v0.2 Inventory - Split Automated testing into separate files - Item tests & inventory tests
